Remove Prettier Formatting and Arrow Funcs
