+++ api/cart.js
@@ -1,23 +1,26 @@
-import { productConstructors, variationChoices } from './products'
-import { getRandomInt } from '../Utilities/math'
+import { productConstructors, variationChoices } from './products';
+import { getRandomInt } from '../Utilities/math';
 
-const getRandomItem = () => {
-    return productConstructors.bed({
-        size: variationChoices.size[getRandomInt(variationChoices.size.length)],
-        color: variationChoices.color[getRandomInt(variationChoices.color.length)],
-    })
+const bedSizes = variationChoices.size;
+const bedColors = variationChoices.color;
+
+function getRandomItem() {
+  return productConstructors.bed({
+    size: bedSizes.size[getRandomInt(bedSizes.length)],
+    color: bedColors.color[getRandomInt(bedColors.length)],
+  });
 }
 
-export const getCartOfSize = (size) => {
-    let cart = []
-    for (let i = 0; i < size; i++) {
-        cart.push(getRandomItem())
-    }
-    return cart
+export function getCartOfSize(size) {
+  let cart = [];
+  for (let i = 0; i < size; i++) {
+    cart.push(getRandomItem());
+  }
+  return cart;
 }
 
-export const getRandomCart = () => {
-    const rand = getRandomInt(4) + 1
-    const cartSize = rand % 3 === 0 ? 2 : rand % 2 === 0 ? 1 : 0
-    return getCartOfSize(cartSize)
+export function getRandomCart() {
+  const rand = getRandomInt(4) + 1;
+  const cartSize = rand % 3 === 0 ? 2 : rand % 2 === 0 ? 1 : 0;
+  return getCartOfSize(cartSize);
 }
