+++ api/products.js
@@ -1,48 +1,49 @@
-const sizes = ['twin', 'full', 'queen', 'king', 'cali-king']
-const colors = ['walnut', 'natural']
-export const products = ['bed', 'mattress', 'nightstand']
-
-export const productVariation = {
-    bed: ['color', 'size'],
-    mattress: ['size'],
-    nightstand: ['color'],
-}
+const sizes = ['twin', 'full', 'queen', 'king', 'cali-king'];
+const colors = ['walnut', 'natural'];
 
 export const variationChoices = {
-    color: colors,
-    size: sizes,
+  color: colors,
+  size: sizes,
+};
+
+export function makeNightstand({ color }) {
+  return {
+    name: 'nightstand',
+    variation: {
+      color,
+    },
+  };
 }
 
-export const makeNightstand = ({ color }) => {
-    return {
-        name: 'nightstand',
-        variation: {
-            color,
-        },
-    }
+export function makeBed({ color, size }) {
+  return {
+    name: 'bed',
+    variation: {
+      color,
+      size,
+    },
+  };
 }
 
-export const makeBed = ({ color, size }) => {
-    return {
-        name: 'bed',
-        variation: {
-            color,
-            size,
-        },
-    }
+export function makeMattress({ size }) {
+  return {
+    name: 'mattress',
+    variation: {
+      size,
+    },
+  };
 }
 
-export const makeMattress = ({ size }) => {
-    return {
-        name: 'mattress',
-        variation: {
-            size,
-        },
-    }
-}
+export const products = ['bed', 'mattress', 'nightstand'];
 
 export const productConstructors = {
-    bed: makeBed,
-    mattress: makeMattress,
-    nigthstand: makeNightstand,
-}
+  bed: makeBed,
+  mattress: makeMattress,
+  nigthstand: makeNightstand,
+};
+
+export const productVariation = {
+  bed: ['color', 'size'],
+  mattress: ['size'],
+  nightstand: ['color'],
+};
