+++ components/CloseButton.js
@@ -1,11 +1,9 @@
-import React from 'react'
-
 export const CloseButton = ({ onClick, style }) => {
-    return (
-        <button onClick={onClick} style={{ ...style, border: 0, background: 'transparent' }}>
-            <span role="img" aria-label={`remove item link`}>
-                ❌
-            </span>
-        </button>
-    )
-}
+  return (
+    <button onClick={onClick} style={{ ...style, border: 0, background: 'transparent' }}>
+      <span role="img" aria-label={`remove item link`}>
+        ❌
+      </span>
+    </button>
+  );
+};
