+++ hooks/useSessionStorageCart.js
@@ -1,22 +1,22 @@
-import { useState } from 'react'
+import { useState } from 'react';
 
-export const useSessionStorageCart = (initialCart) => {
-    const [cart, innerSetCart] = useState(() => {
-        try {
-            const storedCart = window.sessionStorage.getItem('cart')
-            return storedCart ? JSON.parse(storedCart) : initialCart
-        } catch (error) {
-            return initialCart
-        }
-    })
-
-    function setCart(newCart) {
-        const newCartValue = newCart instanceof Function ? newCart(cart) : newCart
-        innerSetCart(newCartValue)
-        try {
-            window.sessionStorage.setItem('cart', JSON.stringify(newCartValue))
-        } catch (error) {}
+export function useSessionStorageCart(initialCart) {
+  const [cart, innerSetCart] = useState(() => {
+    try {
+      const storedCart = window.sessionStorage.getItem('cart');
+      return storedCart ? JSON.parse(storedCart) : initialCart;
+    } catch (error) {
+      return initialCart;
     }
+  });
 
-    return [cart, setCart]
+  function setCart(newCart) {
+    const newCartValue = newCart instanceof Function ? newCart(cart) : newCart;
+    innerSetCart(newCartValue);
+    try {
+      window.sessionStorage.setItem('cart', JSON.stringify(newCartValue));
+    } catch (error) {}
+  }
+
+  return [cart, setCart];
 }
