+++ package.json
@@ -1,18 +1,18 @@
 {
-    "name": "thuma-code-challenge",
-    "version": "1.0.0",
-    "scripts": {
-        "dev": "next",
-        "build": "next build",
-        "start": "next start",
-        "post-update": "yarn upgrade --latest"
-    },
-    "dependencies": {
-        "next": "latest",
-        "react": "17.0.2",
-        "react-dom": "17.0.2"
-    },
-    "license": "MIT",
-    "keywords": [],
-    "description": ""
+  "name": "thuma-code-challenge",
+  "version": "1.0.0",
+  "scripts": {
+    "dev": "next",
+    "build": "next build",
+    "start": "next start",
+    "post-update": "yarn upgrade --latest"
+  },
+  "dependencies": {
+    "next": "latest",
+    "react": "17.0.2",
+    "react-dom": "17.0.2"
+  },
+  "license": "MIT",
+  "keywords": [],
+  "description": ""
 }
