+++ pages/index.js
@@ -1,5 +1,3 @@
-import React from 'react'
-
 export default function IndexPage() {
-    return <></>
+  return <></>;
 }
