+++ styles/globals.js
@@ -1,19 +1,19 @@
 export const globals = {
-    centerText: { textAlign: 'center' },
-    capitalize: { textTransform: 'capitalize' },
-    spaceEvenly: { display: 'flex', justifyContent: 'space-evenly' },
-    error: { color: 'red', marginTop: '5px' },
-    primaryButton: {
-        borderRadius: '5px',
-        padding: '1em 2em',
-        background: 'rgba(100,100,100,.1)',
-        fontWeight: 'bold',
-        border: '1px solid rgba(0,0,0,.2)',
-        boxShadow: '0px 2px 3px 0px rgba(200,200,200,.5)',
-    },
-    shadow: { boxShadow: '0px 2px 3px 0px rgba(200,200,200,.5)' },
-    pageHeader: {
-        fontWeight: 300,
-        textAlign: 'center',
-    },
-}
+  centerText: { textAlign: 'center' },
+  capitalize: { textTransform: 'capitalize' },
+  spaceEvenly: { display: 'flex', justifyContent: 'space-evenly' },
+  error: { color: 'red', marginTop: '5px' },
+  primaryButton: {
+    borderRadius: '5px',
+    padding: '1em 2em',
+    background: 'rgba(100,100,100,.1)',
+    fontWeight: 'bold',
+    border: '1px solid rgba(0,0,0,.2)',
+    boxShadow: '0px 2px 3px 0px rgba(200,200,200,.5)',
+  },
+  shadow: { boxShadow: '0px 2px 3px 0px rgba(200,200,200,.5)' },
+  pageHeader: {
+    fontWeight: 300,
+    textAlign: 'center',
+  },
+};
